+++ include/modules/MusicCameraLock.hpp
@@ -2,6 +2,7 @@
 
 #include "module/Module.hpp"
 #include <Geode/modify/EditorUI.hpp>
+#include <Geode/modify/LevelEditorLayer.hpp>
 
 class $module(MusicCameraLock) {
     bool m_cameraLocked = false;
@@ -17,5 +18,8 @@ class $modify(MCLEditorUI, EditorUI) {
     void onPlayback(CCObject* sender);
     void onPlaytest(CCObject* sender);
 
+};
+
+class $modify(MCLLevelEditorLayer, LevelEditorLayer) {
     void lockCamera(float dt);
 };